LPD-89104 Move restore-from-trash tests to ObjectEntryLocalServiceTest
The restore tests exercise ObjectEntryLocalService.restoreObjectEntryFromTrash, so they belong in ObjectEntryLocalServiceTest rather than TrashEntryLocalServiceTest. Each public @Test now mirrors the service method and dispatches to private scenario helpers, with @TestInfo on the public test since it is lost on private methods. The shared add-entry and move-to-trash helpers move to a new module-local CMSObjectEntryTestUtil, and the delete test is renamed from testDeleteEntriesEmptiesRecycleBinForGroup to testDeleteEntries.
